+++ src/main/java/com/springhttpserver/Database/ConnectionType.java
@@ -1,6 +1,0 @@
-package com.springhttpserver.Database;
-
-public enum ConnectionType {
-    hibernate,
-    jdbc
-};