+++ src/main/java/com/springhttpserver/Database/DataAccess/IUserRepository.java
@@ -1,0 +1,7 @@
+package com.springhttpserver.Database.DataAccess;
+
+import com.springhttpserver.Models.UserProfile;
+import org.springframework.data.repository.CrudRepository;
+
+public interface IUserRepository extends CrudRepository<UserProfile, String> {
+}