+++ src/main/java/com/springhttpserver/Database/DriverType.java
@@ -1,5 +1,0 @@
-package com.springhttpserver.Database;
-
-public enum DriverType {
-    JDBC
-}