+++ src/main/java/com/springhttpserver/Database/JDBCConnection.java
@@ -1,4 +1,0 @@
-package com.springhttpserver.Database;
-
-public class JDBCConnection {
-}