+++ src/main/java/com/springhttpserver/Services/IUserProfileService.java
@@ -5,5 +5,6 @@ import com.springhttpserver.Models.UserProfile;
 public interface IUserProfileService {
     UserProfile GetUserProfile(String userId);
     UserProfile CreateNewUserProfile(UserProfile userProfile);
-    boolean DeleteUserProfile(String userId);
+    UserProfile UpdateUserProfile(String userId, UserProfile userProfile);
+    void DeleteUserProfile(String userId);
 }