+++ src/main/java/com/springhttpserver/SpringFoxConfig.java
@@ -2,11 +2,14 @@ package com.springhttpserver;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
+import springfox.bean.validators.configuration.BeanValidatorPluginsConfiguration;
 import springfox.documentation.builders.PathSelectors;
 import springfox.documentation.builders.RequestHandlerSelectors;
 import springfox.documentation.service.ApiInfo;
 import springfox.documentation.service.Contact;
 import springfox.documentation.spi.DocumentationType;
+import springfox.documentation.spring.data.rest.configuration.SpringDataRestConfiguration;
 import springfox.documentation.spring.web.plugins.Docket;
 import springfox.documentation.swagger2.annotations.EnableSwagger2;
 
@@ -14,6 +17,7 @@ import java.util.Collections;
 
 @Configuration
 @EnableSwagger2
+@Import({SpringDataRestConfiguration.class, BeanValidatorPluginsConfiguration.class})
 public class SpringFoxConfig {
         @Bean
         public Docket SwaggerAPI() {