+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+#### 15-09-20
+* Commit to using .yml for application properties
+    * This was also causing database connection issue on startup
+* Validation on swagger api
+* Api try-catches
+    * try to return useful errors (not 500)
+* Begin correctly using beans and annotations to achieve spring standards.
+* Implement CrudReposistory to access MySql db
+    * Along with annotations on UserProfile class with auto update database enabled on hibernate
+
 #### 04-09-20
 * More app settings
 * Starting with connection to rds using annotations on objects using ORM
@@ -6,7 +16,7 @@
 
 #### 01-09-20
 * Application properties using .yml file
-* More work on connecting to database
+* More work done on connecting to database
     * Use JDBC to create connection
         * Still want to create neat generic connection functions such as read and write
 * Package rename